requirements.txt
# Core (Cross-Platform)
customtkinter>=5.2.0
pillow>=10.0.0
tkinterdnd2>=0.3.0

# Document
python-docx>=1.0.0
pypdf2>=3.0.0
pdf2image>=1.16.0

# Image (Extended)
pillow-heif>=0.15.0
cairosvg>=2.7.0

# Video/Audio
ffmpeg-python>=0.2.0

# Build
pyinstaller>=6.0.0